+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.0.2
+- Add show cmdbox checkbox for commands.
+- Change it so the output box size is modifyable.
+
 ## 1.0.1
 - Add optional error popups on command fail.
 - Popup appears on top even if the window is not minimized.